# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Apis",
        "description": "A simplest boilerplate for Laravel Apis",
        "contact": {
            "name": "Gilberto Junior",
            "email": "contact@codehappy.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user list.",
                "description": "Paginate all users.",
                "operationId": "paginateUser",
                "parameters": [
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Search by user's name.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[email]",
                        "in": "query",
                        "description": "Search by user's email.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort by.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "name",
                                "-name",
                                "created_at",
                                "-created_at",
                                "updated_at",
                                "-updated_at"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Get users by page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Show x users per page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#components/schemas/User"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "pagination": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#components/schemas/User"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User doesn't have the right permissions.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Forbidden."
                                }
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a new user.",
                "description": "Create a new user with permissions.",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Form request.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User doesn't have the right permissions.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Forbidden."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "field1": [
                                            "The field1 field is required."
                                        ],
                                        "field2": [
                                            "The selected field2 is invalid."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{uuid}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get the user data.",
                "description": "Display the user data.",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The UUID of the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "75a0f2d3-a0d3-4b0b-b02f-65e70656361d"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User doesn't have the right permissions.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Forbidden."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Not found."
                                }
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update the user data.",
                "description": "Update the user with permissions.",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The UUID of the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "75a0f2d3-a0d3-4b0b-b02f-65e70656361d"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Form request.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "": {
                                                    "$ref": "#components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User doesn't have the right permissions.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Forbidden."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "field1": [
                                            "The field1 field is required."
                                        ],
                                        "field2": [
                                            "The selected field2 is invalid."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete the user data.",
                "description": "Remove the user. The permissions related with the user won't be removed.",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The UUID of the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "75a0f2d3-a0d3-4b0b-b02f-65e70656361d"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted."
                    },
                    "401": {
                        "description": "User not authenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User doesn't have the right permissions.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Forbidden."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Not found."
                                }
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "JsonResponse": {
                "description": "The structure of the api response.",
                "properties": {
                    "message": {
                        "description": "The message of the response.",
                        "type": "string"
                    },
                    "errors": {
                        "description": "The error messages of the response if exists.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
            "UserRequest": {
                "properties": {
                    "name": {
                        "description": "The name of the user.",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email of the user.",
                        "type": "string"
                    },
                    "roles": {
                        "description": "The set of role names.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "permissions": {
                        "description": "The set of permission names.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object",
                "example": {
                    "name": "John Doe",
                    "email": "john@exampple.com",
                    "roles": [
                        "Administrator"
                    ],
                    "permissions": [
                        "custom.permission1",
                        "custom.permission2"
                    ]
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}